+++ CHANGELOG.md
@@ -5,6 +5,20 @@ CHANGELOG
 
 _(None)_
 
+## 2.11.0 (2020-09-30)
+
+- Do not oversimplify types for display when running an update or preview.
+  [#5440](https://github.com/pulumi/pulumi/pull/5440)
+
+- Pulumi Windows CLI now uploads all VCS information to console
+  (fixes [#5014](https://github.com/pulumi/pulumi/issues/5014))
+  [#5406](https://github.com/pulumi/pulumi/pull/5406)
+
+- .NET SDK: Support `Output<object>` for resource output properties
+  (fixes [#5446](https://github.com/pulumi/pulumi/issues/5446))
+  [#5465](https://github.com/pulumi/pulumi/pull/5465)
+
+
 ## 2.10.2 (2020-09-21)
 
 - [sdk/go] Add missing Version field to invokeOptions
@@ -12,7 +26,7 @@ _(None)_
 
 - Add `pulumi console` command which opens the currently selected stack in the Pulumi console.
   [#5368](https://github.com/pulumi/pulumi/pull/5368)
-
+  
 - Python SDK: Cast numbers intended to be integers to `int`.
   [#5419](https://github.com/pulumi/pulumi/pull/5419)